[add] SetsuTexts Container Component
